# lib/api-zod/src/generated/types/galleryImage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gouda Giggles Charcuterie API
 * OpenAPI spec version: 0.1.0
 */

export interface GalleryImage {
  id: number;
  url: string;
  caption: string;
  category: string;
}