# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gouda Giggles Charcuterie API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';

export interface Order {
  id: number;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  eventDate: string;
  deliveryAddress: string;
  /** @nullable */
  specialInstructions?: string | null;
  status: string;
  total: number;
  items: OrderItem[];
  createdAt: string;
}