# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gouda Giggles Charcuterie API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';

export interface OrderInput {
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  eventDate: string;
  deliveryAddress: string;
  /** @nullable */
  specialInstructions?: string | null;
  items: OrderItem[];
}