# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gouda Giggles Charcuterie API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  category: string;
  /** @nullable */
  serves?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  inStock: boolean;
  featured?: boolean;
}