# lib/api-zod/src/generated/types/quote.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gouda Giggles Charcuterie API
 * OpenAPI spec version: 0.1.0
 */

export interface Quote {
  id: number;
  name: string;
  email: string;
  phone: string;
  eventType: string;
  eventDate: string;
  guestCount: number;
  message: string;
  createdAt: string;
}